Refactored to package for readability
