+++ UoitDCLibraryBooking/src/androidTest/java/com/objectivetruth/uoitlibrarybooking/ActivityAboutMeTests/AboutMeDrawerInstrumentationTest.java
@@ -5,7 +5,7 @@ import android.support.test.espresso.matcher.ViewMatchers;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.test.suitebuilder.annotation.LargeTest;
-import com.objectivetruth.uoitlibrarybooking.ActivityAboutMe;
+import com.objectivetruth.uoitlibrarybooking.about.ActivityAboutMe;
 import com.objectivetruth.uoitlibrarybooking.R;
 import com.squareup.spoon.Spoon;
 import org.junit.After;
