+++ UoitDCLibraryBooking/src/androidTest/java/com/objectivetruth/uoitlibrarybooking/ActivityGuidelinesAndPoliciesTests/GuidelinesPoliciesDrawerInstrumentationTest.java
@@ -4,7 +4,7 @@ import android.support.test.espresso.contrib.DrawerActions;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.test.suitebuilder.annotation.LargeTest;
-import com.objectivetruth.uoitlibrarybooking.GuidelinesPoliciesActivity;
+import com.objectivetruth.uoitlibrarybooking.guidelinespolicies.GuidelinesPoliciesActivity;
 import com.objectivetruth.uoitlibrarybooking.R;
 import com.squareup.spoon.Spoon;
 import org.junit.After;
