+++ UoitDCLibraryBooking/src/main/java/com/objectivetruth/uoitlibrarybooking/calendar/whatsnew/WhatsNewDialog.java
@@ -1,8 +1,9 @@
-package com.objectivetruth.uoitlibrarybooking;
+package com.objectivetruth.uoitlibrarybooking.calendar.whatsnew;
 
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
+import com.objectivetruth.uoitlibrarybooking.R;
 import com.objectivetruth.uoitlibrarybooking.Utilities.ResourceLoadingUtilities;
 
 class WhatsNewDialog {
