+++ UoitDCLibraryBooking/src/main/res/layout/activity_about.xml
@@ -24,7 +24,7 @@ android:layout_height="match_parent"
 
             >
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:id="@+id/textView1"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
@@ -46,7 +46,7 @@ android:layout_height="match_parent"
                 android:layout_alignParentRight="true"
                 android:text="Disable Ads (1$)" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 			        android:id="@+id/textView2"
                     android:layout_marginTop="20dp"
 			        android:layout_width="wrap_content"
