+++ UoitDCLibraryBooking/src/main/res/layout/calendar_table_item_actual.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:orientation="vertical" >
 
-    <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+    <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
         android:id="@android:id/text1"
         android:gravity="center"
         android:layout_width="match_parent"
