+++ UoitDCLibraryBooking/src/main/res/layout/diafrag_generic.xml
@@ -8,7 +8,7 @@
         android:layout_width="275dp"
         android:layout_margin="20dp"
         android:layout_height="wrap_content">
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/title_generic_diafrag"
 
             android:layout_width="wrap_content"
@@ -20,7 +20,7 @@
             android:textColor="@color/blue_font"
             android:textAppearance="?android:attr/textAppearanceLarge"
             android:textSize="@dimen/dialog_title_text" />
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:layout_marginTop="22dp"
             android:id="@+id/description_generic_diafrag"
             android:layout_width="match_parent"
