+++ UoitDCLibraryBooking/src/main/res/layout/guidelines_policies.xml
@@ -18,7 +18,7 @@ android:layout_height="match_parent"
 		    android:layout_width="wrap_content"
             android:padding="20dp"
 		    android:layout_height="wrap_content">
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 
                 android:id="@+id/guidelines_policies_landing_title_word1"
                 android:layout_width="wrap_content"
@@ -32,7 +32,7 @@ android:layout_height="match_parent"
                 android:textAppearance="?android:attr/textAppearanceLarge"
                 android:textSize="@dimen/dialog_title_text" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 	            android:id="@+id/guidelines_policies_landing_para1"
 	            android:textColor="@android:color/black"
 	            android:layout_width="wrap_content"
@@ -45,7 +45,7 @@ android:layout_height="match_parent"
                 android:layout_marginBottom="20dp"
 	            android:textAppearance="?android:attr/textAppearanceMedium" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 	            android:id="@+id/guidelines_policies_landing_title_word2"
 	            android:layout_width="wrap_content"
 	            android:layout_below="@+id/guidelines_policies_landing_para1"
@@ -58,7 +58,7 @@ android:layout_height="match_parent"
                 android:textSize="@dimen/dialog_title_text"
 	            android:textColor="@color/blue_font" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 	            android:id="@+id/guidelines_policies_landing_para2"
 	            android:layout_width="wrap_content"
 	            android:layout_height="wrap_content"
