+++ UoitDCLibraryBooking/src/main/res/layout/interaction_joinorleave.xml
@@ -38,7 +38,7 @@
         android:layout_height="wrap_content"
         android:background="@color/white"
         >
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 
             android:id="@+id/joinorleave_date"
             android:layout_width="wrap_content"
@@ -63,7 +63,7 @@
             android:scaleType="centerCrop"
             android:src="@drawable/unknown_room" />
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:layout_marginTop="10dp"
             android:id="@+id/joinorleave_error_text"
             android:visibility="invisible"
@@ -80,7 +80,7 @@
 
 
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/textView1"
             android:layout_below="@+id/room_landing_room_picture"
             android:layout_width="wrap_content"
@@ -92,7 +92,7 @@
             android:text="Room: "
             android:textAppearance="?android:attr/textAppearanceMedium" />
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/joinorleave_room_number"
             android:layout_width="wrap_content"
             android:layout_alignTop="@+id/textView1"
