+++ UoitDCLibraryBooking/src/main/res/layout/interaction_viewleaveorjoin.xml
@@ -13,7 +13,7 @@
         android:layout_height="wrap_content"
         android:background="@android:color/background_light"
         android:orientation="vertical" >
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 
             android:id="@+id/viewleaveorjoin_date"
             android:layout_width="wrap_content"
@@ -38,7 +38,7 @@
             android:scaleType="centerCrop"
             android:src="@drawable/unknown_room" />
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:layout_marginTop="10dp"
             android:id="@+id/viewleaveorjoin_error_text"
             android:visibility="invisible"
@@ -52,7 +52,7 @@
             android:gravity="center_horizontal"
             android:text="@string/error_room_no_exist_comment"
             />
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/room_word"
             android:layout_below="@+id/room_landing_room_picture"
             android:layout_width="wrap_content"
@@ -64,7 +64,7 @@
             android:text="Room: "
             android:textAppearance="?android:attr/textAppearanceMedium" />
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/viewleaveorjoin_room"
             android:layout_width="wrap_content"
             android:layout_alignTop="@+id/room_word"
@@ -89,7 +89,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="10dp" >
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:id="@+id/textView3"
                     android:layout_width="0dp"
                     android:layout_weight="1"
@@ -98,7 +98,7 @@
                     android:text="Group Name:"
                     android:textAppearance="?android:attr/textAppearanceLarge" />
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:id="@+id/viewleaveorjoin_groupname"
                     android:layout_width="0dp"
                     android:layout_weight="1"
@@ -115,7 +115,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="10dp">
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:layout_width="0dp"
                     android:layout_weight="1"
                     android:gravity="right"
@@ -123,7 +123,7 @@
                     android:text="Group Code:"
                     android:textAppearance="?android:attr/textAppearanceMedium" />
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:id="@+id/viewleaveorjoin_groupcode"
                     android:layout_width="0dp"
                     android:layout_weight="2"
@@ -139,7 +139,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="10dp">
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 
                     android:layout_width="0dp"
                     android:layout_weight="1"
@@ -148,7 +148,7 @@
                     android:text="Time:"
                     android:textAppearance="?android:attr/textAppearanceMedium" />
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:id="@+id/viewleaveorjoin_timerange"
                     android:layout_width="0dp"
                     android:layout_weight="2"
@@ -164,7 +164,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="10dp">
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
 
                     android:layout_width="0dp"
                     android:layout_weight="1"
@@ -173,7 +173,7 @@
                     android:text="Institution:"
                     android:textAppearance="?android:attr/textAppearanceMedium" />
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:id="@+id/viewleaveorjoin_institution"
                     android:layout_width="0dp"
                     android:layout_weight="2"
@@ -189,7 +189,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginBottom="10dp">
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:layout_width="0dp"
                     android:layout_weight="1"
                     android:gravity="right"
@@ -197,7 +197,7 @@
                     android:text="Notes: "
                     android:textAppearance="?android:attr/textAppearanceMedium" />
 
-                <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+                <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                     android:id="@+id/viewleaveorjoin_notes"
                     android:layout_width="0dp"
                     android:layout_weight="2"
