+++ UoitDCLibraryBooking/src/main/res/layout/item_table1.xml
@@ -6,7 +6,7 @@
     android:gravity="center"
     android:orientation="vertical" >
 
-    <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+    <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
         android:id="@android:id/text1"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
