+++ UoitDCLibraryBooking/src/main/res/layout/my_bookings_item.xml
@@ -4,7 +4,7 @@
     android:layout_height="wrap_content"
     android:orientation="vertical" >
 
-    <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+    <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
         android:id="@+id/my_booking_item_color"
         android:textColor="@color/blue_font"
         android:layout_width="match_parent"
