+++ UoitDCLibraryBooking/src/main/res/layout/my_bookings_login.xml
@@ -7,7 +7,7 @@
     android:background="@drawable/dialog"
     android:layout_height="match_parent" >
 
-    <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+    <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
         android:layout_centerHorizontal="true"
         android:id="@+id/error_textview"
         android:layout_width="wrap_content"
