+++ UoitDCLibraryBooking/src/main/res/layout/my_bookings_tab.xml
@@ -11,7 +11,7 @@
         android:layout_marginBottom="10dp"
         android:layout_marginTop="10dp" >
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/book_room_actual"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -21,7 +21,7 @@
             
             android:textAppearance="?android:attr/textAppearanceMedium" />
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/joinorleave_room_number"
             android:layout_width="0dp"
             android:layout_weight="1"
@@ -31,7 +31,7 @@
             android:gravity="center_horizontal"
             android:textAppearance="?android:attr/textAppearanceMedium" />
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/book_room_number_actual"
             android:layout_width="0dp"
