+++ UoitDCLibraryBooking/src/main/res/layout/premium_promo.xml
@@ -5,7 +5,7 @@
     
     android:orientation="vertical" >
 
-    <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+    <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
         android:id="@+id/textviewwords"
         android:text="@string/why_upgrade"
         android:layout_width="wrap_content"
