+++ UoitDCLibraryBooking/src/main/res/layout/room_landing.xml
@@ -13,7 +13,7 @@
         android:orientation="vertical"
         >
 
-        <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+        <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
             android:id="@+id/room_landing_top_title"
             android:layout_width="wrap_content"
             android:layout_height="match_parent"
@@ -41,7 +41,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content" >
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_floor_word"
                 android:layout_width="0dp"
                 android:layout_weight="1"
@@ -50,7 +50,7 @@
                 android:text="Floor:"
                 android:textAppearance="?android:attr/textAppearanceMedium" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_floor_number"
                 android:layout_width="0dp"
                 android:layout_weight="2"
@@ -68,7 +68,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content" >
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_seating_cap_word"
                 android:layout_width="0dp"
                 android:gravity="right"
@@ -77,7 +77,7 @@
                 android:text="Seating Cap:"
                 android:textAppearance="?android:attr/textAppearanceMedium" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_seating_capacity"
                 android:layout_width="0dp"
                 android:layout_weight="2"
@@ -93,7 +93,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content" >
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_min_bookers_word"
                 android:layout_width="0dp"
                 android:gravity="right"
@@ -102,7 +102,7 @@
                 android:text="Min. Bookers:"
                 android:textAppearance="?android:attr/textAppearanceMedium" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_min_bookers"
                 android:layout_width="0dp"
                 android:layout_weight="2"
@@ -118,7 +118,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content" >
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_max_bookable_word"
                 android:layout_width="0dp"
                 android:layout_weight="1"
@@ -127,7 +127,7 @@
                 android:text="Max. Time:"
                 android:textAppearance="?android:attr/textAppearanceMedium" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_max_time"
                 android:layout_width="0dp"
                 android:layout_weight="2"
@@ -143,7 +143,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content" >
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_comments_word"
                 android:layout_width="0dp"
                 android:layout_weight="1"
@@ -152,7 +152,7 @@
                 android:text="Facilities: "
                 android:textAppearance="?android:attr/textAppearanceMedium" />
 
-            <com.objectivetruth.uoitlibrarybooking.RobotoTextView
+            <com.objectivetruth.uoitlibrarybooking.app.userinterfaceUI.RobotoTextView
                 android:id="@+id/room_landing_comments"
                 android:layout_width="0dp"
                 android:layout_weight="2"
